+++ packages/ag-charts-types/src/main.ts
@@ -23,7 +23,8 @@ export * from './chart/types';
 export * from './chart/zoomOptions';
 export * from './chartBuilderOptions';
 export * from './presets/financial/financialOptions';
-export * from './presets/gauge/gaugeCommonOptions';
+export * from './presets/financial/priceVolumeOptions';
+export * from './presets/gauge/commonOptions';
 export * from './presets/gauge/gaugeOptions';
 export * from './presets/gauge/linearGaugeOptions';
 export * from './presets/gauge/radialGaugeOptions';
