+++ packages/ag-charts-types/src/presets/gauge/commonOptions.ts
