+++ packages/ag-charts-types/src/presets/gauge/gaugeOptions.ts
@@ -1,4 +1,4 @@
 import type { AgLinearGaugePreset } from './linearGaugeOptions';
 import type { AgRadialGaugePreset } from './radialGaugeOptions';
 
-export type AgGaugePreset = AgLinearGaugePreset | AgRadialGaugePreset;
+export type AgGaugePresets = AgLinearGaugePreset | AgRadialGaugePreset;
